HW done

+++ src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -61,7 +61,7 @@ public class StudentController {
     }
 
     @GetMapping("formula")
-    public ResponseEntity<Integer> formula() {
+    public ResponseEntity<Long> formula() {
         return ResponseEntity.ok(service.formula());
     }
